Feature/shellcode (#10)
* Initial implementation

* Make some refactoring for future implementation

* Refactored more

* UI works like it should

* Bugfix

* Bugfix on Linux

+++ src/orthia/orthia_lib/orthia_files_win32.h
@@ -270,7 +270,7 @@ int LoadFileToVector_Silent(const std::wstring& fileName,
         return 0;
 
     char* pData = &data[0];
-    return file.ExactRead_Silent(pData, (ULONG)fileSize);
+    return file.ExactRead_Silent(pData, (unsigned int)fileSize);
 }
 
 bool IsFileExists(const std::wstring& fileName);
